# backend/client/utente-client/fascicolo/anagrafica/api-doc.json
{
	"swagger": "2.0",
	"info": {
		"description": "Modulo per l'erogazione dei servizi di anagrafica del fascicolo",
		"version": "1.0.0",
		"title": "A4G - Fascicolo - Anagrafica"
	},
	"host": "itts55.intra.infotn.it:18080",
	"basePath": "/anagrafica",
...
	"paths": {
...
		"/api/v1/fascicolo": {
			"get": {
				"tags": [
					"fascicolo-controller"
				],
				"summary": "Ricerca paginata dei fascicolo secondo i criteri di ricerca impostati",
				"operationId": "ricercaFascicoloUsingGET",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "cuaa",
						"in": "query",
						"description": "CUAA delle istruttorie da ricercare",
						"required": false,
						"type": "string"
					},
					{
						"name": "numeroElementiPagina",
						"in": "query",
						"description": "Numero elementi per pagina. Valore di default: 10",
						"required": false,
						"type": "integer",
						"format": "int32"
					},
					{
						"name": "ordine",
						"in": "query",
						"description": "Direzione dell'ordinamento (ASC o DESC). Valore di default: ASC",
						"required": false,
						"type": "string",
						"enum": [
							"ASC",
							"DESC"
						]
					},
					{
						"name": "pagina",
						"in": "query",
						"description": "Numero di pagina. Valore di default: 0",
						"required": false,
						"type": "integer",
						"format": "int32"
					},
					{
						"name": "proprieta",
						"in": "query",
						"description": "Proprieta' dell'ordinamento (Default: id)",
						"required": false,
						"type": "string"
					},
					{
						"name": "ragioneSociale",
						"in": "query",
						"description": "Ragione sociale/denominazione delle istruttorie da ricercare",
						"required": false,
						"type": "string"
					},
					{
						"name": "revocaOrdinariaMandato",
						"in": "query",
						"description": "Fascicoli che hanno una Revoca Ordinaria. Data di Revoca Ordinaria",
						"required": false,
						"type": "boolean"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/Rappresenta l'oggetto contenente i risultati di una ricerca paginata«FascicoloDto»"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			}
		},
...
		"/api/v1/fascicolo/legacy": {
			"get": {
				"tags": [
					"fascicolo-ags-controller"
				],
				"summary": "Reperimento informazioni del Fascicolo aziendale tramite id Ags",
				"operationId": "getFascicoloByIdUsingGET",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "id",
						"in": "query",
						"description": "ID ags del fascicolo",
						"required": true,
						"type": "integer",
						"format": "int64",
						"allowEmptyValue": false
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/FascicoloAgsDto"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			}
		},
		"/api/v1/fascicolo/legacy/paged": {
			"get": {
				"tags": [
					"fascicolo-ags-controller"
				],
				"summary": "Reperimento informazioni paginate del Fascicolo aziendale tramite condizioni in OR",
				"operationId": "getFascicoliUsingGET",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "cuaa",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "denominazione",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "numeroElementiPagina",
						"in": "query",
						"description": "Numero elementi per pagina. Valore di default: 10",
						"required": false,
						"type": "integer",
						"format": "int32"
					},
					{
						"name": "ordine",
						"in": "query",
						"description": "Direzione dell'ordinamento (ASC o DESC). Valore di default: ASC",
						"required": false,
						"type": "string",
						"enum": [
							"ASC",
							"DESC"
						]
					},
					{
						"name": "pagina",
						"in": "query",
						"description": "Numero di pagina. Valore di default: 0",
						"required": false,
						"type": "integer",
						"format": "int32"
					},
					{
						"name": "proprieta",
						"in": "query",
						"description": "Proprieta' dell'ordinamento (Default: id)",
						"required": false,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/Rappresenta l'oggetto contenente i risultati di una ricerca paginata«FascicoloAgsDto»"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			}
		},
		"/api/v1/fascicolo/legacy/{cuaa}": {
			"get": {
				"tags": [
					"fascicolo-ags-controller"
				],
				"summary": "Reperimento informazioni Fascicolo aziendale",
				"operationId": "getFascicoloUsingGET",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "cuaa",
						"in": "path",
						"description": "CUAA del fascicolo",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/FascicoloAgsDto"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			}
		},
		"/api/v1/fascicolo/legacy/{cuaa}/soggetti": {
			"get": {
				"tags": [
					"fascicolo-ags-controller"
				],
				"summary": "Recupera i Soggetti registrati con ‘ruolo esteso’ di Titolare/Rappresentante Legale nel fascicolo aziendale ",
				"operationId": "getSoggettiFascicoloAziendaleUsingGET",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "cuaa",
						"in": "path",
						"description": "CUAA azienda",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/CaricaAgsDto"
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			}
		},
		"/api/v1/fascicolo/legacy/{id}/movimenti-validazione": {
			"get": {
				"tags": [
					"fascicolo-ags-controller"
				],
				"summary": "Recupera il numero di movimentazioni di validazione terminate con successo e la data dell'ultima validazione positiva per l'anno di campagna indicato",
				"operationId": "getMovimentiValidazioneFascicoloUsingGET",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "campagna",
						"in": "query",
						"description": "Anno Campagna",
						"required": true,
						"type": "integer",
						"format": "int64",
						"allowEmptyValue": false
					},
					{
						"name": "id",
						"in": "path",
						"description": "ID del fascicolo",
						"required": true,
						"type": "integer",
						"format": "int64"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/MovimentoValidazioneFascicoloAgsDto"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			}
		},
...
		"/api/v1/fascicolo/private/{cuaa}/controllo-completezza": {
			"get": {
				"tags": [
					"fascicolo-private-controller"
				],
				"summary": "Restituisce l'esito per ogni controllo di completezza del fascicolo. I controlli danno esituo TRUE se superati con successo, FALSE altrimenti.",
				"operationId": "getControlloCompletezzaFascicoloUsingGET",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "cuaa",
						"in": "path",
						"description": "Cuaa azienda per cui si vuole ottenere il report",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "boolean"
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			}
		},
...
		"/api/v1/fascicolo/{cuaa}": {
			"get": {
				"tags": [
					"fascicolo-controller"
				],
				"summary": "Recupero fascicolo per CUAA azienda agricola",
				"operationId": "getByCuaaUsingGET",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "cuaa",
						"in": "path",
						"description": "CUAA azienda agricola",
						"required": true,
						"type": "string"
					},
					{
						"name": "idValidazione",
						"in": "query",
						"description": "idValidazione",
						"required": false,
						"type": "integer",
						"format": "int32"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/FascicoloDto"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			}
		},
...
		"/api/v1/fascicolo/{cuaa}/check-lettura-fascicolo": {
			"get": {
				"tags": [
					"fascicolo-controller"
				],
				"summary": "Controllo se l'utente autenticato ha i permessi sul fascicolo specificato da cuaa",
				"operationId": "checkLetturaFascicoloUsingGET",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "cuaa",
						"in": "path",
						"description": "CUAA azienda agricola",
						"required": true,
						"type": "string"
					},
					{
						"name": "idValidazione",
						"in": "query",
						"description": "idValidazione",
						"required": false,
						"type": "integer",
						"format": "int32"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "boolean"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			}
		},
...
	},
	"definitions": {
...
		"CaricaAgsDto": {
			"type": "object",
			"properties": {
				"carica": {
					"type": "string",
					"description": "Carica ricoperta dalla persona fisica nella persona identificata dal cuaa",
					"enum": [
						"ACCOMANDATARIO_DI_SAPA",
						"AMMINISTRATORE",
						"AMMINISTRATORE_CON_REQUISITI",
						"AMMINISTRATORE_DELEGATO",
						"AMMINISTRATORE_DELEGATO_E_PREPOSTO",
						"AMMINISTRATORE_E_PREPOSTO",
						"AMMINISTRATORE_E_RESPONSABILE_TECNICO",
						"AMMINISTRATORE_GIUDIZIARIO",
						"AMMINISTRATORE_PROVVISORIO",
						"AMMINISTRATORE_STRAORDINARIO",
						"AMMINISTRATORE_UNICO",
						"AMMINISTRATORE_UNICO_E_PREPOSTO",
						"COAMMINISTRATORE",
						"COMMISSARIO_GIUDIZIALE",
						"COMMISSARIO_GIUDIZIARIO",
						"COMMISSARIO_GOVERNATIVO",
						"COMMISSARIO_LIQUIDATORE",
						"COMMISSARIO_MINISTERIALE",
						"COMMISSARIO_PREFETTIZIO",
						"COMMISSARIO_STRAORDINARIO",
						"CONSIGLIERE_DELEGATO",
						"CONSIGLIERE_DELEGATO_E_PREPOSTO",
						"CONSIGLIERE_DELEGATO_E_RESPONSABILE_TECNICO",
						"CONSIGLIERE_E_LEGALE_RAPPRESENTANTE",
						"CURATORE",
						"CURATORE_DELLO_EMANCIPATO",
						"CURATORE_FALLIMENTARE",
						"CURATORE_SPECIALE_DI_MINORE",
						"INSTITORE",
						"LEGALE_RAPPRESENTANTE",
						"LEGALE_RAPPRESENTANTE_FIRMATARIO",
						"LEGALE_RAPPRESENTANTE_ART_2_L_25_8_91_N_287",
						"LEGALE_RAPPRESENTANTE_DI_CUI_ALL_ART_2_DELLA_LEGGE_87_DEL_25_8_1991",
						"LEGALE_RAPPRESENTANTE_DI_INCAPACE",
						"LEGALE_RAPPRESENTANTE_DI_SOCIETA",
						"LEGALE_RAPPRESENTANTE_E_RESPONSABILE_TECNICO",
						"LEGALE_RAPPRESENTANTE_FIGLIO_MINORE",
						"LEGALE_RAPPRESENTANTE_INTESTATARIO_DEL_TESSERINO",
						"LIQUIDATORE",
						"LIQUIDATORE_GIUDIZIARIO",
						"MADRE_ESERCENTE_LA_PATRIA_POTESTA",
						"PADRE_ESERCENTE_LA_PATRIA_POTESTA",
						"PADRE_O_MADRE_ESERCENTE_LA_PATRIA_POTESTA",
						"PREPOSTO_DELLA_SEDE_SECONDARIA",
						"PRESIDENTE",
						"PRESIDENTE_COMITATO_DIRETTIVO",
						"PRESIDENTE_COMITATO_ESECUTIVO",
						"PRESIDENTE_CONSIGLIO_AMMINISTRAZIONE",
						"PRESIDENTE_CONSIGLIO_DIRETTIVO",
						"PRESIDENTE_CONSORZIO",
						"PRESIDENTE_DEL_COMITATO_DI_GESTIONE",
						"PRESIDENTE_DEL_CONSIGLIO_DI_GESTIONE",
						"PRESIDENTE_DI_CONSORZIO",
						"PRESIDENTE_E_AMMINISTRATORE_DELEGATO",
						"PRESIDENTE_E_CONSIGLIERE_DELEGATO",
						"PRESIDENTE_E_PREPOSTO",
						"PRESIDENTE_E_RESPONSABILE_TECNICO",
						"PRESIDENTE_EFFETTIVO_CONSIGLIO_DIRETTIVO",
						"PRESIDENTE_GIUNTA_ESECUTIVA",
						"PROCURATORE_GENERALE",
						"PROPRIETARIO",
						"PROPRIETARIO_AUTORIZZATO_A_RISCUOTERE_E_QUIETANZARE",
						"RAPPRESENTANTE_IN_ITALIA",
						"RAPPRESENTANTE_LEGALE_DELLE_SEDI_SECONDARIE",
						"RAPPRESENTANTE_LEGALE_DI_CUI_ALL_ART_2_LEGGE_REG_N_37_DEL_30_8_1988",
						"RAPPRESENTANTE_LEGALE_DI_CUI_ALL_ART_93_DEL_RD_18_6_1931_N_773",
						"RAPPRESENTANTE_PREPOSTO_A_SEDE_SECONDARIA_IN_ITALIA",
						"RAPPRESENTANTE_PREPOSTO_ALLA_DIPENDENZA_IN_ITALIA",
						"RAPPRESENTANTE_SOCIETA_ESTERA",
						"RAPPRESENTANTE_STABILE",
						"RAPPRESENTATE_LEGALE_DI_CUI_ALL_ART_93_DEL_RD_18_6_1931_N_773",
						"SOCIO",
						"SOCIO_ACCOMANDATARIO",
						"SOCIO_ACCOMANDATARIO_D_OPERA",
						"SOCIO_ACCOMANDATARIO_E_PREPOSTO",
						"SOCIO_ACCOMANDATARIO_E_RAPPRESENTANTE_LEGALE",
						"SOCIO_AMMINISTRATORE",
						"SOCIO_CON_FIRMA_CONGIUNTA",
						"SOCIO_CONTITOLARE",
						"SOCIO_DI_SOCIETA_DI_PERSONE_RAPPRES",
						"SOCIO_E_LEGALE_RAPPRESENTANTE",
						"SOCIO_RAPPRESENTANTE",
						"SOCIO_UNICO",
						"TITOLARE",
						"TITOLARE_DELL_IMPRESA_ARTIGIANA",
						"TITOLARE_DI_CUI_ART_2_LEGGE_25_8_91_N_287",
						"TITOLARE_E_RESPONSABILE_TECNICO",
						"TITOLARE_FIRMATARIO",
						"TUTORE",
						"VICE_PRESIDENTE_E_CONSIGLIERE_DELEGATO"
					]
				},
				"codiceFiscale": {
					"type": "string",
					"description": "Codice fiscale di una persona fisica"
				},
				"cognome": {
					"type": "string",
					"description": "cognome della persona"
				},
				"cuaa": {
					"type": "string",
					"description": "Cuaa della pesona in cui la persona fisica ha la carica"
				},
				"denominazione": {
					"type": "string",
					"description": "Denominazione della persona"
				},
				"nome": {
					"type": "string",
					"description": "nome della persona"
				}
			},
			"title": "CaricaAgsDto",
			"description": "Rappresenta una carica che una persona fisica ha in una persona giuridica o in una ditta individuale"
		},
...
		"DetenzioneDto": {
			"type": "object",
			"properties": {
				"codiceFiscale": {
					"type": "string",
					"description": "Codice fiscale"
				},
				"dataFine": {
					"type": "string",
					"format": "date",
					"description": "Dati fine detenzione"
				},
				"dataInizio": {
					"type": "string",
					"format": "date",
					"description": "Data inizio detenzione"
				},
				"denominazione": {
					"type": "string",
					"description": "Denominazione fascicolo"
				},
				"id": {
					"type": "integer",
					"format": "int64",
					"description": "Id detenzione"
				},
				"tipoDetenzione": {
					"type": "string",
					"description": "Tipo di detenzione del fascicolo (e.g. 'mandato' oppure 'detenzione in proprio')",
					"enum": [
						"MANDATO",
						"DETENZIONE_IN_PROPRIO"
					]
				}
			},
			"title": "DetenzioneDto",
			"description": "Detenzione di un fascicolo "
		},
...
		"FascicoloAgsDto": {
			"type": "object",
			"properties": {
				"caa": {
					"type": "string"
				},
				"cuaa": {
					"type": "string"
				},
				"dataAggiornamento": {
					"type": "string",
					"format": "date-time"
				},
				"dataCostituzione": {
					"type": "string",
					"format": "date-time"
				},
				"dataValidazione": {
					"type": "string",
					"format": "date-time"
				},
				"denominazione": {
					"type": "string"
				},
				"idAgs": {
					"type": "integer",
					"format": "int64"
				},
				"identificativoSportello": {
					"type": "integer",
					"format": "int64"
				},
				"iscrittoSezioneSpecialeAgricola": {
					"type": "boolean"
				},
				"nonIscrittoSezioneSpecialeAgricola": {
					"type": "boolean"
				},
				"organismoPagatore": {
					"type": "string"
				},
				"pec": {
					"type": "string"
				},
				"sportello": {
					"type": "string"
				},
				"stato": {
					"type": "string",
					"enum": [
						"IN_LAVORAZIONE",
						"IN_ANOMALIA",
						"CESSATO",
						"CHIUSO",
						"VALIDO"
					]
				}
			},
			"title": "FascicoloAgsDto"
		},
...
		"FascicoloDto": {
			"type": "object",
			"properties": {
				"cameraCommercioCodiceFiscaleDifferente": {
					"type": "boolean"
				},
				"cuaa": {
					"type": "string",
					"description": "Codice fiscale azienda agricola"
				},
				"dataApertura": {
					"type": "string",
					"format": "date",
					"description": "Data apertura del fascicolo"
				},
				"dataModifica": {
					"type": "string",
					"format": "date",
					"description": "Data ultima modifica del fascicolo"
				},
				"dataValidazione": {
					"type": "string",
					"format": "date",
					"description": "Data validazione del fascicolo"
				},
				"denominazione": {
					"type": "string",
					"description": "Denominazione Impresa"
				},
				"detenzione": {
					"description": "Dati del Mandato di un fascicolo ",
					"$ref": "#/definitions/DetenzioneDto"
				},
				"id": {
					"type": "integer",
					"format": "int64",
					"description": "Identificativo fascicolo"
				},
				"organismoPagatore": {
					"type": "string",
					"description": "Stato del fascicolo",
					"enum": [
						"APPAG",
						"ALTRO_OP"
					]
				},
				"stato": {
					"type": "string",
					"description": "Stato del fascicolo",
					"enum": [
						"IN_AGGIORNAMENTO",
						"ALLA_FIRMA_AZIENDA",
						"VALIDATO",
						"IN_VALIDAZIONE",
						"ALLA_FIRMA_CAA",
						"FIRMATO_CAA",
						"IN_ATTESA_TRASFERIMENTO",
						"TRASFERITO"
					]
				}
			},
			"title": "FascicoloDto",
			"description": "Dati anagrafici generali del fascicolo "
		},
...
		"MovimentoValidazioneFascicoloAgsDto": {
			"type": "object",
			"properties": {
				"cuaa": {
					"type": "string"
				},
				"dataUltimaValidazionePositiva": {
					"type": "string",
					"format": "date-time"
				},
				"denominazione": {
					"type": "string"
				},
				"idFascicolo": {
					"type": "integer",
					"format": "int64"
				},
				"validazioniPositive": {
					"type": "integer",
					"format": "int32"
				}
			},
			"title": "MovimentoValidazioneFascicoloAgsDto"
		},
...
	}
}